Organized Components, change groupBy Keys in data-context
