+++ kimchechallenge/src/components/ShowCountries/ShowCountries.jsx
@@ -6,7 +6,7 @@ import {
     AccordionItemButton,
     AccordionItemPanel,
 } from 'react-accessible-accordion';
-import { CountryName, Group, GroupsContainer, GroupTitle } from '../styledComponents/styledGroup'
+import { CountryName, Group, GroupsContainer, GroupTitle } from '../../styledComponents/styledGroup'
 import CountryPropsList from './CountryPropsList';
 
 const ShowCountries = ({ filteredCountries }) => {
